refactor(ri): colocate stories alongside component implementations
Move 10 story files and shared mock from top-level stories/ directory
into their respective component directories. Update Storybook config
to remove the old stories/ glob, and simplify jest testMatch to a
single colocated pattern.
