+++ packages/reference-implementation/src/components/CredentialInfo/CredentialInfo.stories.tsx
@@ -1,17 +1,17 @@
 import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
-import { JsonBlock } from '../src/components/JsonBlock';
-import { storyCredentialMock } from './StoryVCMock';
+import { CredentialInfo } from '../CredentialInfo';
+import { storyCredentialMock } from '../__mocks__/StoryVCMock';
 
 const meta = {
-  title: 'Json Block',
-  component: JsonBlock,
+  title: 'Credential Info',
+  component: CredentialInfo,
   parameters: { layout: 'centered' },
   tags: ['autodocs'],
-} satisfies Meta<typeof JsonBlock>;
+} satisfies Meta<typeof CredentialInfo>;
 
 export default meta;
-type Story = StoryObj<typeof meta>;
+type Story = StoryObj<typeof CredentialInfo>;
 
 export const Default: Story = {
   args: { credential: storyCredentialMock },
