+++ packages/reference-implementation/src/components/MessageText/MessageText.stories.tsx
@@ -1,0 +1,22 @@
+import React from 'react';
+import type { Meta, StoryObj } from '@storybook/react';
+import { MessageText } from '../MessageText';
+import { Status } from '@reference-implementation/components';
+
+const meta = {
+  title: 'Message text',
+  component: MessageText,
+  parameters: { layout: 'centered' },
+  tags: ['autodocs'],
+} satisfies Meta<typeof MessageText>;
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const Default: Story = {
+  args: {
+    status: Status.error,
+    text: 'Network error',
+  },
+  decorators: [(Story) => <Story />],
+};
