+++ packages/reference-implementation/src/components/__mocks__/StoryVCMock.ts
